Browser_Use/requirements.txt
playwright>=1.48.0
python-dotenv>=1.0.1
browser-use
